User profile

+++ app.py
@@ -158,6 +158,17 @@ def signup():
                 'valid_card_num': False,
                 'success': False
             })
+        elif not len(str(card_num)) == 16:
+            return jsonify({
+                'valid_name': True,
+                'valid_username': True,
+                'username_taken': False,
+                'valid_email': True,
+                'email_taken': False,
+                'valid_password': True,
+                'valid_card_num': False,
+                'success': False
+            })
 
         # Get Key
         key = Key.query.filter_by(id=1).first().key
@@ -237,7 +248,7 @@ def login(email=None, attempted_pwd=None):
     decrypted_saved_pwd = decrypted_saved_pwd.decode('utf-8')
 
     if decrypted_saved_pwd == attempted_pwd:
-        user=User.query.filter_by(email=email).first()
+        user = User.query.filter_by(email=email).first()
         database_update(True, user.id)
         return jsonify(user_info.serialize())
     else:
@@ -594,6 +605,23 @@ def get_user_subscriptions(user_id=None):
     return jsonify({'subscriptions': [subscription_obj.serialize() for subscription_obj in subscription_obj_list]})
 
 
+#[url]/is_slot_deletable
+@app.route('/is_slot_deletable/user=<user_id>', methods=['GET'])
+def is_slot_deletable(user_id=None):
+    if user_id is not None and user_id.isdigit() and int(user_id) > 0:
+        slots = UserSlots.query.filter_by(user_id=user_id).filter_by(delete_slot=False).all()
+        slot_list = list()
+        for slot in slots:
+            slot_list.append(slot.tv_show_id)
+        if len(slots) > 10:
+            return jsonify({'valid_user_id': True,
+                            'is_slot_deletable': True})
+        return jsonify({'valid_user_id': True,
+                        'is_slot_deletable': False})
+    return jsonify({'valid_user_id': False,
+                    'is_slot_deletable': False})
+
+
 # { user_id: [user_id], slot_id: [slot_id] }
 @app.route('/slot/flag/delete', methods=['PUT'])
 def flag_slot_delete():
@@ -644,58 +672,38 @@ def flag_slot_delete():
         return str(e)
 
 
-# { user_id: [user_id] }
-@app.route('/slot/delete', methods=['DELETE'])
-def delete_slots(func_call=False, user_id=None):
+# [url]/server_update
+@app.route('/database_update', methods=['DELETE'])
+def database_update(func_call=False, user_id=None):
     try:
         if func_call is False:
             data = request.get_json()
             user_id = data['user_id']
 
-        if user_id is None or not str(user_id).isdigit() or user_id <= 0:
-            user = None
-        else:
+        if user_id is not None or isinstance(user_id, int) or user_id > 0:
             user = User.query.filter_by(id=user_id).scalar()
-
-        if user is None or user.sub_date >= (datetime.now() - timedelta(30)).date():
-            return jsonify({'valid_user_id': False,
-                            'success': False})
-
-        if not user.num_slots > 10:
-            return jsonify({'valid_user_id': True,
-                            'success': False})
-
-        # Check Backwards
-        for slot_num in range(user.num_slots, 10, -1):
-            slot = UserSlots.query.filter_by(user_id=user.id).filter_by(slot_num=slot_num).first()
-
-            if slot.delete_slot == 1:
-                UserSlots.query.filter_by(user_id=user.id).filter_by(slot_num=slot_num).delete()
-                user.num_slots -= 1
+            is_deletion_success = True
+            if user is not None:
+                if not delete_expired_movies(user_id):
+                    is_deletion_success = False
+                if not delete_slots(user_id):
+                    is_deletion_success = False
+                if not delete_expired_tv_shows(user_id):
+                    is_deletion_success = False
+                if not is_deletion_success:
+                    return jsonify({'success': True,
+                                    'valid_user_id': True,
+                                    'deletions_success': is_deletion_success})
+                # email user
+                user.sub_date = datetime.now()
                 db.session.commit()
-
-        return jsonify({'valid_user_id': True,
-                        'success': True})
-
-    except Exception as e:
-        return str(e)
-
-
-# { user_id: [user_id] }
-# route to clear all slots
-@app.route('/clear_slots', methods=['PUT'])
-def clear_slots():
-    try:
-        data = request.get_json()
-        user_id = data['user_id']
-        if user_id is None or not isinstance(user_id, int) or user_id <= 0:
-            return jsonify({'slots_cleared': False})
-        user_slots_list = UserSlots.query.filter_by(user_id=user_id).all()
-        if user_slots_list:
-            for user_slots in user_slots_list:
-                clear_individual_slot(user_slots.user_id, user_slots.slot_num)
-            db.session.commit()
-            return jsonify({'slots_cleared': True})
+                email_sender.subscription_renew_email(user.username, user.email, user.sub_date + timedelta(30))
+                return jsonify({'success': True,
+                                'valid_user_id': True,
+                                'deletions_success':is_deletion_success})
+        return jsonify({'success': False,
+                        'valid_user_id': False,
+                        'deletions_success': is_deletion_success})
     except Exception as e:
         return str(e)
 
@@ -1029,6 +1037,9 @@ def get_user_movie_list(user_id=None):
             image_url = movie.image_url
             rating = rm_entry.user_rating
 
+            if rating is None:
+                rating = 0
+
             rm = RatedMovie(movie_id, title, image_url, rating)
             rated_movies.append(rm)
 
@@ -1091,6 +1102,9 @@ def get_user_movie_rating(user_id=None, movie_id=None):
 
     else:
         entry = UserRatedMovieRel.query.filter_by(user_id=user_id).filter_by(movie_id=movie_id).first()
+        if entry is None:
+            return jsonify({'movie_rating': 0})
+
         return jsonify({'movie_rating': entry.user_rating})
 
 
@@ -1594,29 +1608,6 @@ def display_timeline(user_id=None):
         return jsonify({'timeline': timeline})
 
 
-# [url]/server_update
-@app.route('/database_update', methods=['DELETE'])
-def database_update(func_call=False, user_id=None):
-    try:
-        if func_call is False:
-            data = request.get_json()
-            user_id = data['user_id']
-
-        if user_id is not None or isinstance(user_id, int) or user_id > 0:
-
-            if User.query.filter_by(id=user_id).scalar() is not None:
-                delete_expired_movies(True, user_id)
-                delete_slots(True, user_id)
-                delete_expired_tv_shows(True, user_id)
-                return jsonify({'success': True,
-                                'valid_user_id': True})
-        else:
-            return jsonify({'success': False,
-                            'valid_user_id': False})
-    except Exception as e:
-        return str(e)
-
-
 # add friend friend and friend adds back
 def add_friend(user_id, friend_id):
     # Add friend to database
@@ -1673,38 +1664,63 @@ def get_average_rating(is_tv_show: bool, media_id: int):
 
 
 # checks database if movies are past rented due date and deletes them
-def delete_expired_movies(func_call=False, user_id=None):
+def delete_expired_movies(user_id=None):
     yesterday_datetime = datetime.now() - timedelta(1)
-    users_list = list()
     # ensures list is not empty
-    if func_call is True and user_id is not None:
-        check_not_empty = UserRentedMovies.query.filter_by(user_id=user_id).filter(
-            UserRentedMovies.rent_datetime <= yesterday_datetime)
-        user = User.query.filter_by(id=user_id).first()
-        users_list.append(user)
-        if check_not_empty:
-            UserRentedMovies.query.filter(UserRentedMovies.rent_datetime <= yesterday_datetime).delete()
-            db.session.commit()
-            for user in users_list:
-                email_sender.movie_return_email(user.username, user.email)
+    if user_id is not None and isinstance(user_id, int) and user_id > 0:
+        check_not_none = UserRentedMovies.query.filter_by(user_id=user_id).filter(
+            UserRentedMovies.rent_datetime <= yesterday_datetime).scalar() is not None
+        if check_not_none:
+            UserRentedMovies.query.filter_by(user_id=user_id).filter(UserRentedMovies.rent_datetime <= yesterday_datetime).delete()
+            user = User.query.filter_by(id=user_id).first()
+            email_sender.movie_return_email(user.username, user.email)
+            return True
+        else:
+            return True
+    return False
+
+
+# { user_id: [user_id] }
+def delete_slots(user_id=None):
+    try:
+        if user_id is None or not str(user_id).isdigit() or user_id <= 0:
+            user = None
+        else:
+            user = User.query.filter_by(id=user_id).first()
+
+        if user is None or user.sub_date >= (datetime.now() - timedelta(30)).date():
+            return False
+
+        if not user.num_slots > 10:
+            return False
+
+        # Check Backwards
+        for slot_num in range(user.num_slots, 10, -1):
+            slot = UserSlots.query.filter_by(user_id=user.id).filter_by(slot_num=slot_num).first()
+
+            if slot.delete_slot == 1:
+                UserSlots.query.filter_by(user_id=user.id).filter_by(slot_num=slot_num).delete()
+                user.num_slots -= 1
+
+        return True
+
+    except Exception as e:
+        return str(e)
 
 
 # need to add to login function and need to add check to not allow deleting slots past 10
-def delete_expired_tv_shows(func_call=False, user_id=None):
+def delete_expired_tv_shows(user_id=None):
     month_ago_date = (datetime.now() - timedelta(30)).date()
     # ensures list is not empty
-    if func_call is True and user_id is not None:
-        expired_users = User.query.filter_by(id=user_id).filter(User.sub_date <= month_ago_date)
-        if expired_users:
-            for user in expired_users:
-                remove_list = UserSlots.query.filter_by(user_id=user.id).filter_by(unsubscribe=True)
-                user.sub_date = datetime.now()
-                db.session.commit()
-                email_sender.subscription_renew_email(user.username, user.email, user.sub_date + timedelta(30))
-                for tv_show_to_remove in remove_list:
-                    subscribe(user.id, tv_show_to_remove.tv_show_id, True)
-                    remove_tv_show(user.id, tv_show_to_remove.tv_show_id)
-            db.session.commit()
+    if user_id is not None and isinstance(user_id, int) and user_id > 0:
+        expired_user = User.query.filter_by(id=user_id).filter(User.sub_date <= month_ago_date).scalar()
+        if expired_user is not None:
+            tv_show_to_remove_list = UserSlots.query.filter_by(user_id=user_id).filter_by(unsubscribe=True)
+            for tv_show_to_remove in tv_show_to_remove_list:
+                subscribe(user_id, tv_show_to_remove.tv_show_id, True)
+                remove_tv_show(user_id, tv_show_to_remove.tv_show_id)
+        return True
+    return False
 
 
 # function to delete tv_show in slot
